% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferenceSymbols.R
\name{getReferenceSymbols}
\alias{getReferenceSymbols}
\title{Accessing stock price data}
\usage{
getReferenceSymbols(
  x,
  start_date,
  end_date,
  focus = "cl",
  currency = "USD",
  leverage = 1
)
}
\arguments{
\item{x}{a list of strings containing the stock symbols of the traded asset at focus
as well as stock symbols relevant to the traded asset}

\item{start_date}{a string containing the start date of the data to be collected, expected in yyyy-mm-dd format}

\item{end_date}{a string containing the end date of the data to be collected, expected in yyyy-mm-dd format}

\item{focus}{a string that determines which one value from each stock to export into the resulting xts. \cr
'op'- opening prices, 'hi'- high prices, 'lo'- low prices, 'cl' - closing prices,  'vol'- volume, 'ad'- adjusted closing prices}

\item{currency}{a string containing a currency code. It converts the stock prices into the desired currency}

\item{leverage}{a numerical value by which all asset prices are multiplied with, for example when trading with leverage}
}
\value{
xts file with one price value per day of the traded stock and all reference assets.
}
\description{
Accessing and downloading of stock prices. The package allows an easy way to
download reference stock prices into a xts object. \cr
The resulting object is configured to create compatible input for the \code{\link{walkForwardNeuralnet}} function.
}
\details{
All price data is downloaded from Yahoo Finance.
}
\examples{
getReferenceSymbols(x = c('XOM','BP','COP','OXY','VLO','TOT','SLB'), start_date = '2000-10-10', end_date = '2020-10-10')
}
