% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorChart.R
\name{indicatorChart}
\alias{indicatorChart}
\title{Plotting financial indicator charts}
\usage{
indicatorChart(indicator, symbol = NULL, zoom = NULL)
}
\arguments{
\item{indicator}{an xts/zoo object of a time series containing singular values for each time interval}

\item{symbol}{a string containing the related stock code of the indicator. Plots the stock price data
aligned to the corresponding indicator values.}

\item{zoom}{a string containing the start and end-point of the enhancement of the chart,
see Details for formatting information.

The correct string format for the zoom is genrally yyyy-mm-dd::yyyy-mm-dd. However it can also be simplified to
yyyy-mm::yyyy-mm, yyyy::yyyy or any combination of those versions.}
}
\description{
Tool to create financial charts of indicators and stock price data.
}
\examples{
indicatorChart(indicator$indicator)
indicatorChart(indicator$indicator, symbol = 'XOM', zoom = '2018-07-01::2019-07-30')
indicatorChart(indicator$indicator, zoom = '2018-07-01::2019-07-30')
indicatorChart(indicator$indicator, symbol = 'XOM')
}
